requirements.txt
slack_bolt>=1.18.0
slack_sdk>=3.21.3
python-dotenv>=1.0.0
Flask>=2.2.5
PyPokerEngine @ git+https://github.com/ishikota/PyPokerEngine.git
